Move settings_helper to gui settings in mantidqtinterfaces
