+++ scripts/Engineering/common/test/test_settings_helper.py
@@ -4,7 +4,7 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-from Engineering.common.settings_helper import set_setting, get_setting
+from mantidqtinterfaces.mantidqtinterfaces.Engineering.gui.engineering_diffraction.settings.settings_helper import set_setting, get_setting
 from qtpy.QtCore import QSettings, QCoreApplication
 
 from shutil import rmtree
